Provide support for shape filling and separate paragraphs handling to separate file
