+++ Word/OoxmlWord.cs
@@ -91,6 +91,17 @@ namespace OpenXML.Words
                 }
             }
         }
+        // Alternative approach if you want to return a copy instead of a readonly wrapper
+        internal List<int> IDs
+        {
+            get { return new List<int>(_IDs); }
+        }
+
+        // Public property for _numberingPart
+        internal NumberingDefinitionsPart NumberingPart
+        {
+            get { return _numberingPart; }
+        }
 
         /**
         #region Create Core Properties for OpenXML Word Document
@@ -190,7 +201,7 @@ namespace OpenXML.Words
                     {
                         case FF.Paragraph ffP:
                             {
-                                var para = CreateParagraph(ffP);
+                                var para = OoxmlParagraph.CreateInstance(_IDs,_numberingPart).CreateParagraph(ffP);
                                 _wpBody.InsertBefore(para, sectionProperties);
                                 break;
                             }
@@ -235,327 +246,6 @@ namespace OpenXML.Words
         #endregion
 
         #region Create OpenXML Paragraph
-        internal WP.Paragraph CreateParagraph(FF.Paragraph ffP)
-        {
-            lock (_lockObject)
-            {
-                try
-                {
-                    var wpParagraph = new WP.Paragraph();
-
-                    if (ffP.Style != null)
-                    {
-                        var paragraphProperties = new WP.ParagraphProperties();
-
-                        var paragraphStyleId = new WP.ParagraphStyleId { Val = ffP.Style };
-                        paragraphProperties.Append(paragraphStyleId);
-
-                        #region Create List Paragraph
-
-                        if (ffP.Style == "ListParagraph")
-                        {
-
-                            // Check if NumberingId already exists
-
-                            var isExist = false;
-                            if (_IDs != null)
-                            {
-                                foreach (var id in _IDs)
-                                {
-                                    if (id == ffP.NumberingId)
-                                    {
-                                        isExist = true;
-
-                                        var numbering = _numberingPart.Numbering;
-                                        var abstractNum = numbering.Elements<WP.AbstractNum>().FirstOrDefault(an => an.AbstractNumberId == ffP.NumberingId);
-                                        if (abstractNum != null)
-                                        {
-                                            var level = abstractNum.Elements<WP.Level>().FirstOrDefault(l => l.LevelIndex == ffP.NumberingLevel - 1);
-                                            if (level != null)
-                                            {
-                                                if (ffP.IsAlphabeticNumber)
-                                                {
-                                                    level.NumberingFormat.Val = WP.NumberFormatValues.LowerLetter;
-                                                    level.LevelText.Val = string.Format("%{0}.", (int)ffP.NumberingLevel);
-                                                }
-                                                else if (ffP.IsRoman)
-                                                {
-                                                    level.NumberingFormat.Val = WP.NumberFormatValues.LowerRoman;
-                                                    level.LevelText.Val = string.Format("%{0}.", (int)ffP.NumberingLevel);
-                                                }
-                                                else if (ffP.IsBullet)
-                                                {
-                                                    level.NumberingFormat.Val = WP.NumberFormatValues.Bullet;
-                                                    level.LevelText.Val = "o";
-                                                }
-                                                if (ffP.IsNumbered)
-                                                {
-                                                    level.NumberingFormat.Val = WP.NumberFormatValues.Decimal;
-                                                    level.LevelText.Val = string.Format("%{0}.", string.Join(".%", Enumerable.Range(1, (int)ffP.NumberingLevel)));
-                                                }
-                                                numbering.Save();
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                            if (!isExist)
-                            {
-                                if (ffP.NumberingId != null)
-                                {
-                                    if (ffP.NumberingLevel == null)
-                                        ffP.NumberingLevel = 1;
-                                    if (ffP.IsAlphabeticNumber == false && ffP.IsBullet == false &&
-                                        ffP.IsNumbered == false && ffP.IsRoman == false)
-                                        ffP.IsNumbered = true;
-
-                                    var abstractNum = new WP.AbstractNum() { AbstractNumberId = ffP.NumberingId };
-                                    abstractNum.AddNamespaceDeclaration("w", "http://schemas.openxmlformats.org/wordprocessingml/2006/main");
-
-                                    var multiLevelType = new WP.MultiLevelType() { Val = WP.MultiLevelValues.Multilevel };
-                                    multiLevelType.AddNamespaceDeclaration("w", "http://schemas.openxmlformats.org/wordprocessingml/2006/main");
-
-                                    abstractNum.Append(multiLevelType);
-
-                                    var level = new WP.Level() { LevelIndex = 0 };
-
-                                    for (var i = 1; i <= 9; i++)
-                                    {
-                                        level = new WP.Level() { LevelIndex = i - 1 };
-                                        level.AddNamespaceDeclaration("w", "http://schemas.openxmlformats.org/wordprocessingml/2006/main");
-
-                                        var numberingFormat = new WP.NumberingFormat();
-                                        var levelText = new WP.LevelText();
-
-                                        if (ffP.IsNumbered)
-                                        {
-                                            numberingFormat.Val = WP.NumberFormatValues.Decimal;
-                                            levelText.Val = string.Format("%{0}.", string.Join(".%", Enumerable.Range(1, i)));
-                                        }
-                                        else if (ffP.IsAlphabeticNumber)
-                                        {
-                                            numberingFormat.Val = WP.NumberFormatValues.LowerLetter;
-                                            levelText.Val = string.Format("%{0}.", i);
-                                        }
-                                        else if (ffP.IsRoman)
-                                        {
-                                            numberingFormat.Val = WP.NumberFormatValues.LowerRoman;
-                                            levelText.Val = string.Format("%{0}.", i);
-                                        }
-                                        else if (ffP.IsBullet)
-                                        {
-                                            numberingFormat.Val = WP.NumberFormatValues.Bullet;
-                                            levelText.Val = "o";
-                                        }
-
-                                        var previousParagraphProperties = new WP.PreviousParagraphProperties();
-                                        var indentation = new WP.Indentation() { Left = (i * 720).ToString(), Hanging = "360" };
-                                        previousParagraphProperties.Append(indentation);
-
-                                        level.Append(new WP.StartNumberingValue() { Val = 1 });
-                                        level.Append(numberingFormat);
-                                        level.Append(levelText);
-                                        level.Append(new WP.LevelJustification() { Val = WP.LevelJustificationValues.Left });
-                                        level.Append(previousParagraphProperties);
-
-                                        abstractNum.Append(level);
-                                    }
-
-                                    var numberingInstance = new WP.NumberingInstance() { NumberID = ffP.NumberingId };
-                                    var abstractNumId = new WP.AbstractNumId() { Val = ffP.NumberingId };
-
-                                    numberingInstance.Append(abstractNumId);
-                                    _numberingPart.Numbering.Append(abstractNum);
-                                    _numberingPart.Numbering.Append(numberingInstance);
-
-
-                                    _IDs.Add((int)ffP.NumberingId);
-                                }
-                            }
-
-                            var numberingProperties = new WP.NumberingProperties();
-                            var numberingLevelReference = new WP.NumberingLevelReference() { Val = ffP.NumberingLevel - 1 };
-                            var numberingId = new WP.NumberingId() { Val = ffP.NumberingId };
-                            numberingProperties.Append(numberingLevelReference);
-                            numberingProperties.Append(numberingId);
-                            paragraphProperties.Append(numberingProperties);
-                        }
-                        #endregion
-
-
-                        // Create Borders
-                        if (ffP.ParagraphBorder.Size > 0)
-                        {
-                            WP.ParagraphBorders paragraphBorders = new WP.ParagraphBorders();
-                            WP.TopBorder topBorder = new WP.TopBorder()
-                            {
-                                Val = CreateBorder(ffP.ParagraphBorder.Width),
-                                Color = ffP.ParagraphBorder.Color,
-                                Size = (DF.UInt32Value)(uint)ffP.ParagraphBorder.Size,
-                                Space = (DF.UInt32Value)(uint)ffP.ParagraphBorder.Size
-                            };
-                            WP.LeftBorder leftBorder = new WP.LeftBorder()
-                            {
-                                Val = CreateBorder(ffP.ParagraphBorder.Width),
-                                Color = ffP.ParagraphBorder.Color,
-                                Size = (DF.UInt32Value)(uint)ffP.ParagraphBorder.Size,
-                                Space = (DF.UInt32Value)(uint)ffP.ParagraphBorder.Size
-                            };
-                            WP.BottomBorder bottomBorder = new WP.BottomBorder()
-                            {
-                                Val = CreateBorder(ffP.ParagraphBorder.Width),
-                                Color = ffP.ParagraphBorder.Color,
-                                Size = (DF.UInt32Value)(uint)ffP.ParagraphBorder.Size,
-                                Space = (DF.UInt32Value)(uint)ffP.ParagraphBorder.Size
-                            };
-                            WP.RightBorder rightBorder = new WP.RightBorder()
-                            {
-                                Val = CreateBorder(ffP.ParagraphBorder.Width),
-                                Color = ffP.ParagraphBorder.Color,
-                                Size = (DF.UInt32Value)(uint)ffP.ParagraphBorder.Size,
-                                Space = (DF.UInt32Value)(uint)ffP.ParagraphBorder.Size
-                            };
-
-                            paragraphBorders.Append(topBorder);
-                            paragraphBorders.Append(leftBorder);
-                            paragraphBorders.Append(bottomBorder);
-                            paragraphBorders.Append(rightBorder);
-
-                            paragraphProperties.Append(paragraphBorders);
-                        }
-                        // Create Justification
-                        WP.JustificationValues justificationValue = CreateJustification(ffP.Alignment);
-                        paragraphProperties.Append(new WP.Justification { Val = justificationValue });
-
-                        // Create Indentation
-                        CreateIndentation(paragraphProperties, ffP.Indentation);
-
-                        wpParagraph.Append(paragraphProperties);
-                    }
-
-
-                    foreach (var ffR in ffP.Runs)
-                    {
-                        var wpRun = new WP.Run();
-
-                        var runProperties = new WP.RunProperties();
-
-                        if (ffR.FontFamily != null)
-                        {
-                            var runFont = new WP.RunFonts
-                            {
-                                Ascii = ffR.FontFamily,
-                                HighAnsi = ffR.FontFamily,
-                                ComplexScript = ffR.FontFamily,
-                                EastAsia = ffR.FontFamily
-                            };
-                            runProperties.Append(runFont);
-                        }
-
-                        if (ffR.Color != null)
-                        {
-                            var color = new WP.Color { Val = ffR.Color };
-                            runProperties.Append(color);
-                        }
-
-                        if (ffR.FontSize > 0)
-                        {
-                            var fontSize = new WP.FontSize { Val = (ffR.FontSize * 2).ToString() };
-                            runProperties.Append(fontSize);
-                        }
-
-                        if (ffR.Bold)
-                        {
-
-                            runProperties.Append(new WP.Bold() { Val = new DF.OnOffValue(true) });
-                        }
-
-                        if (ffR.Italic)
-                        {
-                            runProperties.Append(new WP.Italic());
-                        }
-
-                        if (ffR.Underline)
-                        {
-                            var underline = new WP.Underline { Val = WP.UnderlineValues.Single };
-                            runProperties.Append(underline);
-                        }
-
-                        var text = new WP.Text(ffR.Text);
-                        wpRun.Append(runProperties, text);
-                        wpParagraph.AppendChild(wpRun);
-                    }
-
-                    return wpParagraph;
-                }
-                catch (Exception ex)
-                {
-                    var errorMessage = OWD.OoxmlDocData.ConstructMessage(ex, "Create Paragraph");
-                    throw new Openize.Words.OpenizeException(errorMessage, ex);
-                }
-            }
-        }
-
-        private WP.JustificationValues CreateJustification(FF.ParagraphAlignment alignment)
-        {
-            switch (alignment)
-            {
-                case FF.ParagraphAlignment.Left:
-                    return WP.JustificationValues.Left;
-                case FF.ParagraphAlignment.Center:
-                    return WP.JustificationValues.Center;
-                case FF.ParagraphAlignment.Right:
-                    return WP.JustificationValues.Right;
-                case FF.ParagraphAlignment.Justify:
-                    return WP.JustificationValues.Both;
-                default:
-                    return WP.JustificationValues.Left;
-            }
-        }
-
-        private WP.BorderValues CreateBorder(FF.BorderWidth borderWidth)
-        {
-            switch (borderWidth)
-            {
-                case FF.BorderWidth.Single:
-                    return WP.BorderValues.Single;
-                case FF.BorderWidth.Double:
-                    return WP.BorderValues.Double;
-                case FF.BorderWidth.Dotted:
-                    return WP.BorderValues.Dotted;
-                case FF.BorderWidth.DotDash:
-                    return WP.BorderValues.DotDash;
-                default:
-                    return WP.BorderValues.Single;
-            }
-        }
-
-        private void CreateIndentation(WP.ParagraphProperties paragraphProperties, FF.Indentation ffIndentation)
-        {
-            var indentation = new WP.Indentation();
-
-            if (ffIndentation.Left > 0)
-            {
-                indentation.Left = (ffIndentation.Left * 1440).ToString();
-            }
-
-            if (ffIndentation.Right > 0)
-            {
-                indentation.Right = (ffIndentation.Right * 1440).ToString();
-            }
-
-            if (ffIndentation.FirstLine > 0)
-            {
-                indentation.FirstLine = (ffIndentation.FirstLine * 1440).ToString();
-            }
-
-            if (ffIndentation.Hanging > 0)
-            {
-                indentation.Hanging = (ffIndentation.Hanging * 1440).ToString();
-            }
-
-            paragraphProperties.Append(indentation);
-        }
         #endregion
 
         #region Create OpenXML Table
@@ -594,7 +284,8 @@ namespace OpenXML.Words
                             var ffCell = ffTable.Rows[i].Cells[j];
                             foreach (var ffPara in ffCell.Paragraphs)
                             {
-                                wpCell.Append(CreateParagraph(ffPara));
+                                //CreateParagraph(ffPara));
+                                wpCell.Append(OoxmlParagraph.CreateInstance(_IDs, _numberingPart).CreateParagraph(ffPara));
                             }
 
                             wpRow.Append(wpCell);
@@ -783,10 +474,28 @@ namespace OpenXML.Words
                     var adjustValueList = new A.AdjustValueList();
 
                     presetGeometry.Append(adjustValueList);
-                    var outline = new A.Outline();
+                    var outline = new A.Outline();                    
 
                     shapeProperties.Append(transform2D);
                     shapeProperties.Append(presetGeometry);
+                    switch (shape.FillType)
+                    {
+                        case FF.ShapeFillType.Solid:
+                            A.SolidFill solidFill = new A.SolidFill();
+                            A.RgbColorModelHex rgbColor = new A.RgbColorModelHex()
+                            { Val = shape.FillColors.Color1 };
+                            solidFill.Append(rgbColor);
+                            shapeProperties.Append(solidFill);
+                            break;
+                        case FF.ShapeFillType.Gradient:
+                            shapeProperties.Append(CreateGradientFill(
+                                shape.FillColors.Color1, shape.FillColors.Color2));
+                            break;
+                        case FF.ShapeFillType.Pattern:
+                            shapeProperties.Append(CreatePatternFill(
+                                shape.FillColors.Color1, shape.FillColors.Color2));
+                            break;
+                    }
                     shapeProperties.Append(outline);
 
                     var shapeStyle = new DWS.ShapeStyle();
@@ -859,6 +568,54 @@ namespace OpenXML.Words
                     throw new Openize.Words.OpenizeException(errorMessage, ex);
                 }
             }
+        }
+
+        private static A.GradientFill CreateGradientFill(string color1, string color2)
+        {
+            A.GradientFill gradientFill = new A.GradientFill();
+            A.GradientStopList gradientStopList = new A.GradientStopList();
+
+            // Create start position stop (0%)
+            A.GradientStop gradientStop1 = new A.GradientStop() { Position = 0 };
+            //A.RgbColorModelHex rgbColor1 = new A.RgbColorModelHex() { Val = "FF0000" };
+            A.RgbColorModelHex rgbColor1 = new A.RgbColorModelHex() { Val = color1 };
+            gradientStop1.Append(rgbColor1);
+
+            // Create end position stop (100%)
+            A.GradientStop gradientStop2 = new A.GradientStop() { Position = 100000 };
+            //A.RgbColorModelHex rgbColor2 = new A.RgbColorModelHex() { Val = "0000FF" };
+            A.RgbColorModelHex rgbColor2 = new A.RgbColorModelHex() { Val = color2 };
+            gradientStop2.Append(rgbColor2);
+
+            gradientStopList.Append(gradientStop1);
+            gradientStopList.Append(gradientStop2);
+
+            A.LinearGradientFill linearGradientFill = new A.LinearGradientFill() { Angle = 5400000 }; // 90 degrees
+
+            gradientFill.Append(gradientStopList);
+            gradientFill.Append(linearGradientFill);
+
+            return gradientFill;
+        }
+
+        private static A.PatternFill CreatePatternFill(string color1, string color2)
+        {
+            A.PatternFill patternFill = new A.PatternFill() { Preset = A.PresetPatternValues.SmallGrid };
+
+            A.ForegroundColor fgColor = new A.ForegroundColor();
+            //A.RgbColorModelHex fgRgbColor = new A.RgbColorModelHex() { Val = "FF0000" };
+            A.RgbColorModelHex fgRgbColor = new A.RgbColorModelHex() { Val = color1 };
+            fgColor.Append(fgRgbColor);
+
+            A.BackgroundColor bgColor = new A.BackgroundColor();
+            //A.RgbColorModelHex bgRgbColor = new A.RgbColorModelHex() { Val = "FFFFFF" };
+            A.RgbColorModelHex bgRgbColor = new A.RgbColorModelHex() { Val = color2 };
+            bgColor.Append(bgRgbColor);
+
+            patternFill.Append(fgColor);
+            patternFill.Append(bgColor);
+
+            return patternFill;
         }
 
         private A.ShapeTypeValues CreateShapeType(FF.ShapeType shapeType)
@@ -986,11 +743,11 @@ namespace OpenXML.Words
                     /******************* shapes ****************/
                     var wordprocessingShape01 = CreatePartialShape(
                         shape1.ElementId, shape1.X, shape1.Y,
-                        shape1.Width, shape1.Height, CreateShapeType(shape1.Type)
+                        shape1.Width, shape1.Height, CreateShapeType(shape1.Type),shape1
                         );
                     var wordprocessingShape02 = CreatePartialShape(
                         shape2.ElementId, shape2.X, shape2.Y,
-                        shape2.Width, shape2.Height, CreateShapeType(shape2.Type)
+                        shape2.Width, shape2.Height, CreateShapeType(shape2.Type),shape2
                         );
 
                     /**************** connector *****************/
@@ -1126,7 +883,9 @@ namespace OpenXML.Words
             }
         }
 
-        private DWS.WordprocessingShape CreatePartialShape(int Id, int X, int Y, int Width, int Height, A.ShapeTypeValues shapeTypeValues)
+        private DWS.WordprocessingShape CreatePartialShape(int Id, int X, int Y, int Width, int Height,
+            A.ShapeTypeValues shapeTypeValues,
+            FF.Shape shape)
         {
             lock (_lockObject)
             {
@@ -1165,6 +924,24 @@ namespace OpenXML.Words
 
                     shapeProperties.Append(transform2D);
                     shapeProperties.Append(presetGeometry);
+                    switch (shape.FillType)
+                    {
+                        case FF.ShapeFillType.Solid:
+                            A.SolidFill solidFill = new A.SolidFill();
+                            A.RgbColorModelHex rgbColor = new A.RgbColorModelHex()
+                            { Val = shape.FillColors.Color1 };
+                            solidFill.Append(rgbColor);
+                            shapeProperties.Append(solidFill);
+                            break;
+                        case FF.ShapeFillType.Gradient:
+                            shapeProperties.Append(CreateGradientFill(
+                                shape.FillColors.Color1, shape.FillColors.Color2));
+                            break;
+                        case FF.ShapeFillType.Pattern:
+                            shapeProperties.Append(CreatePatternFill(
+                                shape.FillColors.Color1, shape.FillColors.Color2));
+                            break;
+                    }
                     shapeProperties.Append(outline);
 
                     var shapeStyle = new DWS.ShapeStyle();
@@ -1302,7 +1079,7 @@ namespace OpenXML.Words
 
                                     if (!drawingFound)
                                     {
-                                        elements.Add(LoadParagraph(wpPara, sequence));
+                                        elements.Add(OoxmlParagraph.CreateInstance(_IDs, _numberingPart).LoadParagraph(wpPara, sequence));
                                         sequence++;
                                     }
                                     break;
@@ -1413,212 +1190,6 @@ namespace OpenXML.Words
 
 
         #region Load OpenXML Paragraph
-        internal FF.Paragraph LoadParagraph(WP.Paragraph wpPara, int id)
-        {
-            lock (_lockObject)
-            {
-                try
-                {
-                    var ffP = new FF.Paragraph { ElementId = id };
-
-                    var paraProps = wpPara.GetFirstChild<WP.ParagraphProperties>();
-                    if (paraProps != null)
-                    {
-                        var paraStyleId = paraProps.Elements<WP.ParagraphStyleId>().FirstOrDefault();
-                        if (paraStyleId != null)
-                        {
-                            if (paraStyleId.Val != null) ffP.Style = paraStyleId.Val.Value;
-                        }
-                    }
-
-                    if (ffP.Style == "ListParagraph")
-                    {
-                        if (isNumbered(paraProps))
-                        {
-                            if (_numberingPart != null)
-                            {
-                                if (paraProps.NumberingProperties.NumberingId.Val != null &&
-                                paraProps.NumberingProperties.NumberingLevelReference.Val != null)
-                                {
-                                    ffP.NumberingId = paraProps.NumberingProperties.NumberingId.Val;
-                                    ffP.NumberingLevel = paraProps.NumberingProperties.NumberingLevelReference.Val + 1;
-
-                                    var numbering = _numberingPart.Numbering;
-                                    var abstractNum = numbering.Elements<WP.AbstractNum>().FirstOrDefault(an => an.AbstractNumberId == ffP.NumberingId);
-                                    if (abstractNum != null)
-                                    {
-                                        var level = abstractNum.Elements<WP.Level>().FirstOrDefault(l => l.LevelIndex == ffP.NumberingLevel - 1);
-                                        if (level != null)
-                                        {
-                                            if (level.NumberingFormat.Val == WP.NumberFormatValues.Decimal)
-                                                ffP.IsNumbered = true;
-                                            else if (level.NumberingFormat.Val == WP.NumberFormatValues.LowerLetter)
-                                                ffP.IsAlphabeticNumber = true;
-                                            else if (level.NumberingFormat.Val == WP.NumberFormatValues.LowerRoman)
-                                                ffP.IsRoman = true;
-                                            else if (level.NumberingFormat.Val == WP.NumberFormatValues.Bullet)
-                                                ffP.IsBullet = true;
-                                            else
-                                                ffP.IsNumbered = true;
-                                        }
-                                    }
-                                }
-
-                            }
-                        }
-                    }
-
-                    // Load Border
-                    if (isBordered(paraProps))
-                    {
-                        var topBorder = paraProps.ParagraphBorders?.TopBorder;
-                        if (topBorder != null)
-                        {
-                            ffP.ParagraphBorder.Width = LoadBorder(topBorder.Val);
-                            ffP.ParagraphBorder.Color = topBorder.Color;
-                            ffP.ParagraphBorder.Size = (int)(uint)topBorder.Size;
-                        }
-                    }
-
-                    // Load Justification
-                    if (isJustified(paraProps))
-                    {
-                        var justificationElement = paraProps.Elements<WP.Justification>().FirstOrDefault();
-                        if (justificationElement != null)
-                            ffP.Alignment = LoadAlignment(justificationElement.Val);
-                    }
-                    else ffP.Alignment = FF.ParagraphAlignment.Left;
-
-                    // Load Indentation
-                    if (isIndented(paraProps))
-                    {
-                        var Indentation = paraProps.Elements<WP.Indentation>().FirstOrDefault();
-                        if (Indentation != null)
-                        {
-                            if (Indentation.Left != null)
-                                ffP.Indentation.Left = int.Parse(Indentation.Left);
-                            if (Indentation.Right != null)
-                                ffP.Indentation.Right = int.Parse(Indentation.Right);
-                            if (Indentation.Hanging != null)
-                                ffP.Indentation.Hanging = int.Parse(Indentation.Hanging);
-                            if (Indentation.FirstLine != null)
-                                ffP.Indentation.FirstLine = int.Parse(Indentation.FirstLine);
-                        }
-                    }
-
-                    var runs = wpPara.Elements<WP.Run>();
-
-                    foreach (var wpR in runs)
-                    {
-                        var fontSize = wpR.RunProperties?.FontSize?.Val != null
-                            ? int.Parse(wpR.RunProperties.FontSize.Val)
-                            : (int?)null;
-                        if (fontSize != null) fontSize /= 2;
-                        var ffR = new FF.Run
-                        {
-                            Text = wpR.InnerText,
-                            FontFamily = wpR.RunProperties?.RunFonts?.Ascii ?? null,
-                            FontSize = fontSize ?? 0,
-                            Color = wpR.RunProperties?.Color?.Val ?? null,
-                            Bold = (wpR.RunProperties != null && wpR.RunProperties.Bold != null),
-                            Italic = (wpR.RunProperties != null && wpR.RunProperties.Italic != null),
-                            Underline = (wpR.RunProperties != null && wpR.RunProperties.Underline != null)
-                        };
-                        ffP.AddRun(ffR);
-                    }
-
-                    return ffP;
-                }
-
-                catch (Exception ex)
-                {
-                    var errorMessage = OWD.OoxmlDocData.ConstructMessage(ex, "Load Paragraph");
-                    throw new Openize.Words.OpenizeException(errorMessage, ex);
-                }
-            }
-        }
-
-        private FF.ParagraphAlignment LoadAlignment(WP.JustificationValues justificationValue)
-        {
-            if (justificationValue == WP.JustificationValues.Left)
-                return FF.ParagraphAlignment.Left;
-            else if (justificationValue == WP.JustificationValues.Center)
-                return FF.ParagraphAlignment.Center;
-            else if (justificationValue == WP.JustificationValues.Right)
-                return FF.ParagraphAlignment.Right;
-            else if (justificationValue == WP.JustificationValues.Both)
-                return FF.ParagraphAlignment.Justify;
-            else
-                return FF.ParagraphAlignment.Left;
-        }
-
-        private FF.BorderWidth LoadBorder(WP.BorderValues borderValue)
-        {
-            if (borderValue == WP.BorderValues.Single)
-                return FF.BorderWidth.Single;
-            else if (borderValue == WP.BorderValues.Double)
-                return FF.BorderWidth.Double;
-            else if (borderValue == WP.BorderValues.Dotted)
-                return FF.BorderWidth.Dotted;
-            else if (borderValue == WP.BorderValues.DotDash)
-                return FF.BorderWidth.DotDash;
-            else
-                return FF.BorderWidth.Single;
-        }
-
-        private bool isBordered(WP.ParagraphProperties prop)
-        {
-            try
-            {
-                var paragraphBorders = prop.ParagraphBorders;
-                return true;
-            }
-            catch (Exception ex)
-            {
-                return false;
-            }
-        }
-
-        private bool isJustified(WP.ParagraphProperties prop)
-        {
-            try
-            {
-                var justification = prop.Justification;
-                return true;
-            }
-            catch (Exception ex)
-            {
-                return false;
-            }
-        }
-
-        private bool isIndented(WP.ParagraphProperties prop)
-        {
-            try
-            {
-                var indentation = prop.Indentation;
-                return true;
-            }
-            catch (Exception ex)
-            {
-                return false;
-            }
-        }
-
-        private bool isNumbered(WP.ParagraphProperties prop)
-        {
-            try
-            {
-                var numbering = prop.NumberingProperties;
-                var numberingId = numbering.NumberingId.Val;
-                var numberingRef = numbering.NumberingLevelReference.Val;
-                return true;
-            }
-            catch (Exception ex)
-            {
-                return false;
-            }
-        }
 
         #endregion
 
@@ -1699,7 +1270,6 @@ namespace OpenXML.Words
                     // Determine the shape type
                     var presetGeometry = shapeProperties.GetFirstChild<A.PresetGeometry>();
                     var shapeType = FF.ShapeType.Ellipse; // Default
-
 
                     shapeType = LoadShapeType(presetGeometry.Preset);
 
@@ -1803,7 +1373,7 @@ namespace OpenXML.Words
                             var ffParas = new List<FF.Paragraph>();
                             foreach (var paragraph in wpCell.Elements<WP.Paragraph>())
                             {
-                                ffParas.Add(LoadParagraph(paragraph, 0));
+                                ffParas.Add(OoxmlParagraph.CreateInstance(_IDs, _numberingPart).LoadParagraph(paragraph, 0));
                             }
 
                             var ffCell = new FF.Cell { Paragraphs = ffParas };
